finetune/requirements.txt
faiss-cpu
fastapi==0.115.4
gensim==4.3.3
nltk==3.9.1
numpy==1.26.4
pandas==2.2.3
pydantic==1.10.18
torch
uvicorn==0.15.0
pyarrow==14.0.1
psycopg2-binary==2.9.10
tqdm==4.66.4